pyproject.toml
[tool.poetry]
name = "speechrecognition"
version = "0.1.0"
description = ""
authors = ["ybelmach <yaroslav.belmach.wk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
yt-dlp = "^2024.9.27"
pydub = "^0.25.1"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-multipart = "^0.0.12"
sqlalchemy = "^2.0.35"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
